Add files via upload
Temporarily removed because the webhost that we used is in direct conflict with other modules (spaCy and IBM). As a result, we end up using Selenium to web-scrape the text off of an image uploaded to another web app that took the images from the text (Used the same core we did). However, the methods are still functional!!!